Fixed Issue #6 in MIDI-dot-NET
See:
https://github.com/jstnryan/midi-dot-net/commit/e3c735027cacac34bd75e17864756a029e61464f

+++ MIDI/Win32API.cs
@@ -174,7 +174,7 @@ namespace Midi
         [StructLayout(LayoutKind.Sequential)]
         public struct HMIDIOUT
         {
-            public Int32 handle;
+            public IntPtr handle;
         }
 
         /// <summary>
@@ -183,7 +183,7 @@ namespace Midi
         [StructLayout(LayoutKind.Sequential)]
         public struct HMIDIIN
         {
-            public Int32 handle;
+            public IntPtr handle;
         }
 
         #endregion
